Move to prometheus

+++ pkg/metrics/influx/cpu.go
@@ -22,7 +22,7 @@ const (
 	CPU_COEFF = 1e12
 )
 
-func (flux *Influx) CPUCurrent() (uint64, error) {
+func (flux *Influx) CPUCurrent() (float64, error) {
 	var result, err = flux.Query("SELECT MEAN(value) FROM cpu_usage_total WHERE time > now()-5m")
 	if err != nil {
 		return 0, err
@@ -47,7 +47,7 @@ func (flux *Influx) CPUCurrent() (uint64, error) {
 	}
 	var average, _ = result[0].Series[0].Values[0][1].(json.Number).Float64()
 	average /= CPU_COEFF * flux.CPUFactor() // TODO: remove hardcoded value
-	return uint64(average), nil
+	return average, nil
 }
 
 func (flux *Influx) CPUHistory(from, to time.Time, step time.Duration) (vector.Vec, error) {
